requirements.txt
pycryptodome>=3.14.1
jinja2>=3.1.1